+++ packages/contracts/src/dtos/login.dto.ts
@@ -1,4 +1,3 @@
-import 'reflect-metadata'
 import { Type } from 'class-transformer'
 import { IsEmail, IsNotEmpty, IsString, ValidateNested } from 'class-validator'